Separate and cleanup, need rebuild graph 1, switch to sankey diagram pending
