+++ Homework3/trdli/style.css
@@ -104,6 +104,6 @@ a:link, a:visited, a:hover, a:active {
 }
 
 .fade-in {
-  animation: fadeIn 1s ease-in-out;
+  animation: fadeIn 0.5s ease-in;
   animation-fill-mode: forwards;
 }
